SegmentAnalytics.Analytics — Type

mutable struct Analytics

The type for holding segment configuration and segment state

function Analytics(;options...)

Return instance of Analytics initialized with giving options

Required arguments

write_key::String: Segment write key

Optional arguments

batch_size::Integer: Max size of batch of messages that will be send to Segment. By default equal to 100

on_error::Function: Function that will be executed when there is error from segment API

Arguments for transport configuration

host::String

port::Integer

ssl::Bool

headers::Dict

path::String

retries::Integer

backoff_policy::BackoffPolicy

Example

analytics = SegmentAnalytics.Analytics(
  write_key="write_key",
  batch_size=20,
  on_error=(status, error) -> @error "[status] error",
  host="customhost.com"
)

SegmentAnalytics.track — Function

Tracks an event

Arguments

analytics::Analytics

attrs::Dict: Event payload

Possible attrs

event::String: Event name

properties::Dict: Event properties (optional)

anonymous_id::String: ID for a user when you don't know who they are yet. (optional but you must provide either an anonymous_id or user_id)

context::Dict: (optional)

integrations::Dict: What integrations this event goes to (optional)

message_id::String: ID that uniquely identifies a message across the API. (optional)

timestamp::Union{DateTime, ZonedDateTime}: When the event occurred (optional)

user_id::String: The ID for this user in your database (optional but you must provide either an anonymous_id or user_id)

options::Dict: Options such as user traits (optional)

Example

using SegmentAnalytics

analytics = SegmentAnalytics.Analytics(write_key="write_key")
payload = Dict(
  :event => "Event Name",
  :user_id => "User ID",
  :properties => Dict(:p1 => 1, :p2 => 2),
  :message_id => "custom-message-id"
)

track(analytics, payload)